# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Description": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"Name",
					"Description",
					"amount",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetProduct": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqProduct": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Description": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"Name",
					"Description",
					"amount",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Description": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"Name",
					"Description",
					"amount",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IPRODUCT": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteProduct": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductListReqProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Description": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"Name",
					"Description",
					"amount",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResOrder": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"noOfOrder": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					},
					"noOfPizzas": {
						"type": "number",
						"format": "double"
					},
					"statusOfOrder": {
						"type": "string"
					},
					"bill": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"noOfOrder",
					"productId",
					"noOfPizzas",
					"statusOfOrder",
					"bill"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetOrder": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqOrder": {
				"properties": {
					"noOfOrder": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					},
					"noOfPizzas": {
						"type": "number",
						"format": "double"
					},
					"statusOfOrder": {
						"type": "string"
					},
					"bill": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"noOfOrder",
					"productId",
					"noOfPizzas",
					"statusOfOrder",
					"bill"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqOrder": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"noOfOrder": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					},
					"noOfPizzas": {
						"type": "number",
						"format": "double"
					},
					"statusOfOrder": {
						"type": "string"
					},
					"bill": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"noOfOrder",
					"productId",
					"noOfPizzas",
					"statusOfOrder",
					"bill"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IORDER": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteOrder": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderListReqOrder": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"noOfOrder": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					},
					"noOfPizzas": {
						"type": "number",
						"format": "double"
					},
					"statusOfOrder": {
						"type": "string"
					},
					"bill": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"noOfOrder",
					"productId",
					"noOfPizzas",
					"statusOfOrder",
					"bill"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "Pizza",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/product/getProduct": {
			"post": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetProduct"
							}
						}
					}
				}
			}
		},
		"/product/saveproduct": {
			"post": {
				"operationId": "SaveProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqProduct"
							}
						}
					}
				}
			}
		},
		"/product/updateproduct": {
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqProduct"
							}
						}
					}
				}
			}
		},
		"/product/deleteproduct": {
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Product deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPRODUCT"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteProduct"
							}
						}
					}
				}
			}
		},
		"/product/getproductlist": {
			"get": {
				"operationId": "GetProductList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductListReqProduct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": []
			}
		},
		"/order/getOrder": {
			"post": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetOrder"
							}
						}
					}
				}
			}
		},
		"/order/saveOrder": {
			"post": {
				"operationId": "SaveOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqOrder"
							}
						}
					}
				}
			}
		},
		"/order/updateOrder": {
			"put": {
				"operationId": "UpdateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqOrder"
							}
						}
					}
				}
			}
		},
		"/order/deleteorder": {
			"delete": {
				"operationId": "DeleteOrder",
				"responses": {
					"200": {
						"description": "Order deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORDER"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteOrder"
							}
						}
					}
				}
			}
		},
		"/order/getOrderList": {
			"get": {
				"operationId": "GetOrderList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrderListReqOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}